Advance_Tools_Features/pyproject.toml
[project]
name = "advance-tools-features"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "openai-agents>=0.2.8",
    "tavily-python>=0.7.10",
]
